requirements.txt
git+https://github.com/Hrishi2861/pyrofork-2.2.11-peer-fix.git 
TgCrypto
flask
bs4
python-dotenv
Pillow
praw
aiohttp
redgifs
telegraph